Do not update states after writing cache (#19936)
This is a little preparation for
https://github.com/python/mypy/issues/933.

We need to minimize the number of actions after writing cache, since
these are things workers will need to communicate back to coordinator.
Previously we updated `State.meta` after writing cache, but this looks
not needed anymore, and would complicate parallel checking.

Also delete couple unused attributes on `State`.