[mypyc] Support deleting attributes in __setattr__ wrapper (#19997)
The `__setattr__` wrapper that mypyc generates needs to handle deleting
attributes as well because `del` statements go through the same
`tp_setattro` pointer but with the value argument set to `NULL`.

The wrapper calls `__delattr__` in this case if it's overridden in the
native class (or its parent). Handling of dynamic attributes is
different without `__dict__` which makes a custom `__delattr__` required
if the dynamic attributes are stored in a custom dictionary.

If `__delattr__` is not overridden it calls the implementation of
`object.__delattr__` which results in `AttributeError` because there's
no `__dict__`.

If it's defined without `__setattr__`, mypyc reports an error. It's
possible to support just `__delattr__` but since it shares a slot with
`__setattr__`, the wrapper generation would be more complicated. It
seems like an unlikely use case to only need `__delattr__` so I think it
makes sense to leave it for later.