Fix infinite loop on empty Buffer (#20024)
Currently `Buffer(b"")` (as opposite to `Buffer()`) can go into an
infinite loop when resizing. Fix this by allocating one more byte, that
is conveniently guaranteed by `bytes` ABI.

I also move the `librt/__init__.py` hack to tests, since this is the
only place where it is needed. I am still not sure why `*.so` from
`site-packages` is preferred over a "local" namespace package with the
same `*.so`.