Do not sort unused error codes in unused error codes warning (#20036)
I intuit the previous author of this code sorted the codes for
stability, but it actually should be in default order, to match what the
user typed in. This will be more intuitive for the user.

In my first commit, I add the failing testUnusedIgnoreCodeOrder test. In
my second commit, I fix the code.