[mypyc] Fix inheritance of async defs (#20044)
When inferring a precise generator return type for an async def (or
generator), make the generate returned by an override in a subclass
inherit from the base class generator. This means that the environment
has to be moved to a separate class in the base class generator.

Don't infer a precise generator return type when an override might have
a less precise return type, since it would break LSP.

Fixes mypyc/mypyc#1141.