Make --pretty work better on multi-line issues (#20056)
We can just print the first line for things where there's multiple lines
being the issue. Ideally, we would print the first line, last line, and
explicitly elide the lines in between. That's for a future change!
(maybe)

Fixes https://github.com/python/mypy/issues/18522.

NOTE: this is an aesthetic thing, there's no right formatting. So I
would be fine if this is closed because others think it looks worse.