Switch to a more dynamic SCC processing logic (#20053)
Ref https://github.com/python/mypy/issues/933

Instead of processing SCCs layer by layer, we will now process an SCC as
soon as it is ready. This logic is easier to adapt for parallel
processing, and should get us more benefit from parallelization (as more
SCCs can be processed in parallel). I tried to make order with single
worker stable and very similar (or maybe even identical) to the current
order.

Note I already add some methods to the build manager to emulate parallel
processing, but they are not parallel _yet_.