Allow type parameters reusing the name missing from current module (#20081)
Fixes #18507. Fixes #19526. Fixes #19946 (already closed without
reproducer, but star imports may cause such issues,
`testPEP695TypeVarNameClashStarImport` fails on current master).

Even if a type parameter is reusing the name from outer scope, it still
cannot be a redefinition, so we can safely store it even in presence of
unresolved star imports or variables that do not refer to a type.