Expose buffer cache layout version (#20145)
This idea is quite simple: each time we change how an object is cached
in the buffer we bump this version (even if we don't change the ABI
version). Notes:
* This is *not* the same as ABI version, if the ABI version is
different, the capsule import will fail and you can't even use this
version of `librt`. We can however use buffer cache layout version to
abandon cache files that would result in reading garbage data.
* I can't actually start using this in cache metas in the same PR, to
avoid breaking self-check in CI. If there are no comments objections, I
will merge this soon. Then release a new `librt` version, and make
second PR that actually uses this feature.

While I am at it I fix a bug in the float deserialization (magic value
was not handled).