Do not store deferred NamedTuple fields as redefinitions (#20147)
Fixes #17059.

When we encounter a field with a Placeholder as a default, do not record
it in the new symtable as redefinition - it becomes orphan immediately
and remains there, crashing suring serialization.

The test cases are both crashing on current master. Ideally we hould
emit name-defined in all cases, but it is another unrelated issue
(#17610, and likely some others; this is not specific to named tuples -
plain classes also allow referencing variables that are not defined
yet).