[nit] command_line.rst: the standard format for multiple choice has no space (#19868)
Removing this spaces causes this documentation to match the other
documentation on the page, and also in `--help`

There are no tests for this change. I manually verified that the link to
the option currently is
https://mypy.readthedocs.io/en/stable/command_line.html#cmdoption-mypy-enable-incomplete-feature,
so changing the arguments text does not necessitate adding a new anchor
in to preserve old inbound links to this section.