Update kinds_of_types.rst: keep old anchor for #no-strict-optional (#19828)
Addresses
https://github.com/python/mypy/pull/19252#issuecomment-2953310853
> support this old anchor, for people on older mypy versions

This way, when people get the old documentation link, it will continue
to go to the right place.

I have looked at the documentation that gets generated locally to see if
everything still works. It does. In our vast panoply of documentation,
we now have two very similar refs, `no-strict-optional` and
`no_strict_optional` (used in
https://mypy.readthedocs.io/en/stable/common_issues.html#no-errors-reported-for-obviously-wrong-code
to link to
https://mypy.readthedocs.io/en/stable/command_line.html#no-strict-optional)
but the software handles them fine, without getting confused.