Don't let help formatter line-wrap URLs (#19825)
Fixes #19816

This PR adjusts the `AugmentedHelpFormatter` for the command-line
`--help` output so that it will no longer wrap/break long URLs in the
output.

It also (in a separate commit) adds `https://` in front of two URLs that
lacked it, so they can be recognized as links by anything parsing the
help text.