Do not abort constructing TypeAlias if only type parameters hold us back. (#20162)
Fixes #20135.

I am not fully certain that this is the right way. If the type alias can
be reasonably constructed and only type parameters prevent that, we can
create an "almost" equivalent alias with all placeholder-bearing
typevars replaced with their "simple" equivalents with default
values/bound/default values. This would cause current iteration to not
emit some parameterization errors later, but, since we already defer the
alias as a whole, we will recheck all of them again.

This obviously adds some pointless work (we check parameterization that
will not be used later), but probably is not that big of a deal,
recursive aliases are relatively rare in wild. If this turns out to be a
bottleneck, we can add a `parameters_ready` flag to aliases and skip the
heavy parts if it is False, but I think it isn't necessary now.