Micro-optimize cache primitives in librt.internal (#20194)
This includes the following optimizations:
* Split Buffer into ReadBuffer and WriteBuffer which are more
specialized
* Only check buffer types in wrapper functions, not in C primitives
* Use pointers instead of integer indexes in the buffer objects

This improves the performance of a micro-benchmark that reads integers
in a loop by 5%, but this could help more if we'd inline some of the
smaller functions (in the future). By making the functions simpler,
inlining is more feasible.

I'm not sure what's the best way to merge this -- maybe we'll need to
have broken master for a while, and then we ca publish a new version of
`librt`, and then update mypy to work using the new `librt` version.