Make NoneType annotation error a new code
Change the code of "NoneType should not be used as a type, please use None instead" error to be a new code, so that people who don't like this rule can disable it easily.

I haven't made this a subcode of valid-type because NoneType is actually a valid type; it's just dispreferred.

I have added the documentation for the new code, but haven't added any new tests (the old test still runs fine).

Fixes https://github.com/python/mypy/issues/20218