[librt] Add `librt.internal` API versioning with backward compat (#20221)
This lets us add new API features by adding more functions to the
capsule and the module namespace without breaking backward
compatibility.

We shouldn't use the capsule if the installed version is too old, since
we could be calling functions via uninitialized pointers.

Note that this is a breaking change in the `librt.internal` ABI.