Fix crash on recursive tuple with Hashable (#20232)
Fixes https://github.com/python/mypy/issues/20227

This is huge pain to reproduce, so I added just _some_ repro as a test.
Note the change in `typeops.py` is not required for this fix, but it is
a technically correct thing to do that I noticed while looking at this.