[mypyc] Add minimal, experimental librt.base64 module (#20226)
The module currently only has a `b64encode` function adapted from
CPython. It's only enabled when librt is compiled with experimental
features enabled, so that we are free to iterate on this and break
backward compatibility until we are ready to declare the module as
stable.

This also adds a way to define experimental features in `librt` (and
mypyc in general, but it's currently only used for `librt`).

In follow-up PRs I'm planning to add a more efficient implementation of
`b64encode` and add more features to the module, including decoding. I'm
not planning to include every feature from the stdlib base64 module,
since many of them aren't used very widely.