[mypyc] Add a primitive for bytes "translate" method (#20305)
This only supports one-argument calls. In a microbenchmark the
performance was significantly better for small inputs. It was also
slightly better for long inputs, but only after I added a loop unrolling
hint (on Apple Silicon and AMD + Linux).

I relied on LLM assist as another experiment, but did it in small
increments with detailed instructions.