[mypyc] Wrap async functions with function-like type (#20260)
Currently, introspection functions from the `inspect` module like
`iscoroutinefunction(fn)` don't return the expected result for functions
compiled with mypyc. mypyc defines the functions using `PyMethodDef` in
C for which cpython makes objects with type `PyCFunction_Type` and not
`PyFunction_Type`. The latter corresponds to `types.FunctionType` in
python. So the `isfunction(fn)`
[check](https://github.com/python/cpython/blob/3.14/Lib/inspect.py#L267)
fails.

`iscoroutinefunction(fn)` and some others accept [function-like
objects](https://github.com/python/cpython/blob/3.14/Lib/inspect.py#L2070)
as long as they are callable and have some required attributes to
support compiled functions.

The most important attribute is `__code__` which is a code object that
stores flags for the function. For example for
`iscoroutinefunction(fn)`, the `CO_COROUTINE` flag is
[checked](https://github.com/python/cpython/blob/3.14/Lib/inspect.py#L329).

In this PR, mypyc adds a wrapper `CPyFunction` whose `__call__` calls
the same function that would be defined with `PyMethodDef`. The wrapper
is added to `__dict__` of either the module (for functions) or the
containing class (for methods) with the original name of the function so
the wrapper is returned in place of the `PyCFunction_Type` objects.

For nested functions, mypyc already creates a callable class so the
function-like attributes are added as properties. A static `CPyFunction`
object is created separately for each class that is used in the getter
and setter methods of the callable class.

For now `CPyFunction`s are created only for async functions to support
`iscoroutinefunction(fn)` but the implementation could be extended to
support other introspection functions in future PRs. For non-async
functions the wrapper does not seem necessary for now because the
default behavior makes the function always return false. So the behavior
is as-expected for non-async functions.

This implementation was inspired by
[Cython](https://github.com/cython/cython/blob/master/Cython/Utility/CythonFunction.c)
which generates much richer function wrappers.