[mypyc] lib-rt base64: support pyodide for Python 3.12
For Python 3.12, pyodide requires emscripten 3.1.{46,52,58}, which uses an older
version of SIMDe that doesn't include all the NEON 64-bit intrinsics we need.

So we use the 32-bit NEON intrinsics implementation of base64 for Pyodide on
Python 3.12, and use the 64-bit NEON intrinsics implementations of
base64 for Pyodide on Python 3.13.

Pyodide for Python 3.13 uses emscripten 4.0.{6,8,9}

Emscripten versdion 3.1.63 appears to be the first version with the
updated SIMDe versions:

https://github.com/emscripten-core/emscripten/commit/6223c43516e41d78154f24dd319fc37da3ba1a46