Add hidden --overwrite-union-syntax option (#20332)
Start updating the test output (errors and reveal type) to use the PEP
604 union syntax. The process is similar to what we did last year when
we adopted PEP 585 generics in the test output.

As we still default to Python 3.9 for type checking, this PR adds a
temporary internal flag `--overwrite-union-syntax` to be able to
overwrite the output.

This PR only changes a handful of test files so the modifications to
mypy itself are clearly visible. The goal is to change the test output
for all other tests in followup PRs. Once that's done, I plan to
deprecate the `--force-union-syntax` flag similar to the already
deprecated `--force-uppercase-builtins`.