[stubtest] ignore `__conditional_annotations__` (#20392)
On Python 3.14, this module attribute seems to exist in certain
situations:

```pycon
>>> if True:
...     a: str
...     
>>> __conditional_annotations__
{0}
```

It's a pretty esoteric attribute without any documentation. It seems to
have been added in https://github.com/python/cpython/pull/130935

Either way, stubtest was complaining about it in scipy-stubs. So all
things considered, I figured it'd be best to just ignore it.