[mypyc] Fix calling async methods through vectorcall (#20393)
Fixes https://github.com/mypyc/mypyc/issues/1170

Fixed a bug introduced when adding the async function wrapper in
https://github.com/python/mypy/commit/81eaa5d1fb139d62418e16ee69d45b9c89ce46cb
that would result in `TypeErrors` being raised when the wrapped function
was called using the vectorcall protocol.

The exception was caused by the wrapper keeping the `self` argument of
methods in the `args` vector instead of extracting it to a separate c
function argument. The called function would not expect the `self`
argument to be part of `args` so it appeared as if too many arguments
were passed.