[mypyc] Fix compilation of byteswriter_extra_ops (#20403)
Compiling byteswriter_extra_ops.c on its own fails because it uses types
from Python.h without including it so add the include directive.

It also references `BytesWriterObject` which is defined only when
`MYPYC_EXPERIMENTAL` is defined so wrap the functions with the same
`ifdef`.

Also add the explicit include of Python.h in other header files that
currently rely on the file being included above them when compiling.
Remove unused variables.