Prevent synthetic intersections from leaking to module public interfaces (#20459)
Fixes #20457.

---

In addition, 642cf64 fixes a bug which can be triggered in stubtest.
Reproducer:

```python
# mod.pyi
class A: ...
class B: ...
class AB(A, B): ...

a: A
if isinstance(a, B): ...
```
```python
# mod.py
class A: pass
class B: pass
class AB(A, B): pass

a = A()
```
```shell
$ stubtest mod
error: mod.<subclass of "mod.A" and "mod.B"> is not present at runtime
Stub: in file mod.pyi
<TypeInfo mod.<subclass of "mod.A" and "mod.B">>
Runtime:                                                                                                                                                                                                                                                   
MISSING
```

Admittedly, nobody should be writing `isinstance()` in a stub, but it
doesn't seem to be illegal to do so, and in any case auto-generated
stubs may forget to strip such constructs and surface weird error
messages like the one above.

I tried adding a test to `teststubtest.py`, but it fails because there
doesn't seem to be a way to add fixtures to test cases there, which is
needed to invoke `isinstance()`. The test is

```python
    @collect_cases
    def test_generated_intersection_non_public(self) -> Iterator[Case]:
        yield Case(
            stub="""
class A: ...
class B: ...
class AB(A, B): ...

a: A
if isinstance(a, B): ...
""",
            runtime="""
class A: pass
class B: pass
class AB(A, B): pass

a: A
""",
            error=None,
        )
```